tests: add elem from array tests with resize

+++ src/GF_sum.c
@@ -1,5 +1,5 @@
 #include <assert.h>
-#include <stdint.h>
+#include <stddef.h>
 
 #include "GF.h"
 #include "poly.h"
@@ -14,16 +14,16 @@ int GF_elem_sum(GF_elem_t *res, GF_elem_t a, GF_elem_t b) {
   }
 
   // Dimension of the field extension.
-  int8_t n = a.GF->n;
+  size_t n = a.GF->n;
 
   // Characteristic of the field.
-  int8_t p = h.GF->p;
+  uint8_t p = h.GF->p;
 
   int8_t *w = h.poly->coeff;
   int8_t *u = a.poly->coeff;
   int8_t *v = b.poly->coeff;
 
-  for (int8_t i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     assert(u[i] >= 0);
     assert(v[i] >= 0);
     w[i] = (u[i] + v[i]) % p;
